refactor: index.jsx 파일이 LiveLists 컴포넌트를 불러오는 방식으로 변경
