--- frontend/src/components/LiveLists/LiveList.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const LiveList = () => {};
-
-export default LiveList;
